add HTML

+++ projectmanagement/src/main/java/mom/sda/projectmanagement/entities/ProjectEntity.java
@@ -12,12 +12,12 @@ public class ProjectEntity {
     private String description;
     private String administrator;
 
-    public long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(long id) {
-        this.id = Math.toIntExact(id);
+    public void setId(Integer id) {
+        this.id = id;
     }
 
     public String getName() {
